+++ TranspositionTable.h
@@ -42,7 +42,7 @@ class TranspositionTable {
         };
 
         Entry* table;
-        int32_t bucketSizes[bucketCount];
+        uint8_t bucketSizes[bucketCount];
     
     public:
         TranspositionTable();
